Fix bugs.

+++ back-office-interface/src/main/java/app/bo/api/book/SearchRecordAJAXResponse.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public class SearchRecordAJAXResponse {
     @Property(name = "total")
-    public Integer total;
+    public Long total;
 
     @Property(name = "records")
     public List<BorrowedRecordView> borrowedRecords;
